# Conflicts:
#	src/main/kotlin/engine/physics/PointMass.kt

+++ src/main/kotlin/engine/shape/ShapeQueries.kt
@@ -13,7 +13,6 @@ import kotlin.math.sqrt
  * This keeps shapes runtime-agnostic and ensures broadphase queries can go through one path.
  */
 object ShapeQueries {
-
     /**
      * Builds a world-space AABB by projecting local bounds through [transform].
      */
@@ -22,13 +21,11 @@ object ShapeQueries {
             is CircleShape -> worldAabbForCircle(shape, transform)
             is RectangleShape -> worldAabbForRectangle(shape, transform)
         }
-
     /**
      * Tests a world-space [point] against a local [shape] projected by [transform].
      */
     fun containsPoint(shape: Shape2D, transform: Transform, point: Vector2D): Boolean {
         val localPoint = worldToLocal(point, transform) ?: return false
-
         return when (shape) {
             is CircleShape -> localPoint.lengthSquared() <= shape.radius.toDouble() * shape.radius
             is RectangleShape -> {
@@ -38,11 +35,9 @@ object ShapeQueries {
             }
         }
     }
-
     private fun worldAabbForCircle(shape: CircleShape, transform: Transform): Aabb2D {
         val center = transform.position
         val radius = shape.radius.toDouble()
-
         val scaledX = radius * abs(transform.scale.x)
         val scaledY = radius * abs(transform.scale.y)
 
@@ -54,17 +49,16 @@ object ShapeQueries {
         val extentY = sqrt((scaledX * s) * (scaledX * s) + (scaledY * c) * (scaledY * c))
 
         return Aabb2D(
-            minX = (center.x - extentX).toFloat(),
-            minY = (center.y - extentY).toFloat(),
-            maxX = (center.x + extentX).toFloat(),
-            maxY = (center.y + extentY).toFloat()
+            minX = (center.x - extentX),
+            minY = (center.y - extentY),
+            maxX = (center.x + extentX),
+            maxY = (center.y + extentY)
         )
     }
 
     private fun worldAabbForRectangle(shape: RectangleShape, transform: Transform): Aabb2D {
         val halfWidth = shape.width * 0.5
         val halfHeight = shape.height * 0.5
-
         val corner0 = localToWorld(Vector2D(-halfWidth, -halfHeight), transform)
         val corner1 = localToWorld(Vector2D(halfWidth, -halfHeight), transform)
         val corner2 = localToWorld(Vector2D(halfWidth, halfHeight), transform)
@@ -91,10 +85,10 @@ object ShapeQueries {
         if (corner3.y > maxY) maxY = corner3.y
 
         return Aabb2D(
-            minX = minX.toFloat(),
-            minY = minY.toFloat(),
-            maxX = maxX.toFloat(),
-            maxY = maxY.toFloat()
+            minX = minX,
+            minY = minY,
+            maxX = maxX,
+            maxY = maxY
         )
     }
